+++ zmq/backend/cython/_device.pyx
@@ -27,6 +27,8 @@ from libzmq cimport zmq_device, zmq_proxy, ZMQ_VERSION_MAJOR
 from zmq.backend.cython.socket cimport Socket as cSocket
 from zmq.backend.cython.checkrc cimport _check_rc
 
+from zmq.error import InterruptedSystemCall
+
 #-----------------------------------------------------------------------------
 # Basic device API
 #-----------------------------------------------------------------------------
@@ -52,9 +54,15 @@ def device(int device_type, cSocket frontend, cSocket backend=None):
         return proxy(frontend, backend)
 
     cdef int rc = 0
-    with nogil:
-        rc = zmq_device(device_type, frontend.handle, backend.handle)
-    _check_rc(rc)
+    while True:
+        with nogil:
+            rc = zmq_device(device_type, frontend.handle, backend.handle)
+        try:
+            _check_rc(rc)
+        except InterruptedSystemCall:
+            continue
+        else:
+            break
     return rc
 
 def proxy(cSocket frontend, cSocket backend, cSocket capture=None):
@@ -80,9 +88,15 @@ def proxy(cSocket frontend, cSocket backend, cSocket capture=None):
         capture_handle = capture.handle
     else:
         capture_handle = NULL
-    with nogil:
-        rc = zmq_proxy(frontend.handle, backend.handle, capture_handle)
-    _check_rc(rc)
+    while True:
+        with nogil:
+            rc = zmq_proxy(frontend.handle, backend.handle, capture_handle)
+        try:
+            _check_rc(rc)
+        except InterruptedSystemCall:
+            continue
+        else:
+            break
     return rc
 
 __all__ = ['device', 'proxy']